+++ src/primitives/mod.rs
@@ -1,9 +1,19 @@
+mod aabb2;
+mod aabb3;
 mod point2;
 mod point3;
+mod segment2;
+mod triangle2;
+mod triangle3;
 mod vector2;
 mod vector3;
 
+pub use aabb2::Aabb2;
+pub use aabb3::Aabb3;
 pub use point2::Point2;
 pub use point3::Point3;
+pub use segment2::Segment2;
+pub use triangle2::Triangle2;
+pub use triangle3::Triangle3;
 pub use vector2::Vector2;
 pub use vector3::Vector3;